refactor(core): split model protocol call paths
