+++ packages/core/src/ai-model/service-caller/chat-completion/types.ts
@@ -1,0 +1,33 @@
+import type { StreamingCallback } from '@/types';
+import type OpenAI from 'openai';
+import type { ChatCompletionMessageParam } from 'openai/resources/index';
+import type { ModelRuntime } from '../../models';
+import type { createChatClient } from '../openai-client';
+
+export type ChatCompletionCallOptions = {
+  client: Awaited<ReturnType<typeof createChatClient>>;
+  modelRuntime: ModelRuntime;
+  messages: ChatCompletionMessageParam[];
+  requestConfig: Record<string, unknown>;
+  effectiveTimeoutMs: number | null;
+  abortSignal?: AbortSignal;
+  startTime: number;
+  internalCallId: string;
+};
+
+export type StreamingChatCompletionCallOptions = ChatCompletionCallOptions & {
+  onChunk: StreamingCallback;
+  recordEvent?: (event: Record<string, unknown>) => void;
+};
+
+export type ChatCompletionCallResult = {
+  content: string;
+  accumulatedReasoning: string;
+  rawChoiceMessage?: unknown;
+  usage?: OpenAI.CompletionUsage;
+  timeCost?: number;
+  requestId?: string | null;
+  responseModelName?: string;
+  /** Whether usage for the final returned result has already been reported. */
+  usageReported: boolean;
+};
